feat(addon): wire up worker to calling function

+++ src/addon/utils.h
@@ -1,13 +1,13 @@
 #ifndef BLST_TS_UTILS_H
 #define BLST_TS_UTILS_H
 
-#include "sodium.h"
+#include <sodium.h>
 
 void randomBytesNonZero(uint8_t *out, const size_t length)
 {
     randombytes_buf(out, length);
 
-    for (int i = 0; i < length; i++)
+    for (size_t i = 0; i < length; i++)
     {
         if (out[i] == 0)
         {
